Fix a bug in the script and namelist setting up the test cases
The switch efactor_method in kpp.nml was not correctly updated for
the OSMOSIS cases when setting up the run. It should be set to 4 as
OSMOSIS cases read in surface Stokes drift and decay depth instead of
wave spectrum. This switch is now set in the default kpp.nml for each
test case individually instead of being adjusted according to
turbmethod.