Fix a bug in OSCPapa
Fix a bug in the mean wave direction in the wave spectrum data of OCS
Papa. In the original wave spectrum data in netCDF the mean wave
direction is the "from direction" in degree clockwise from the north
whereas the "to direction" in degree counterclockwise from the east is
used to compute the x- and y-components. A 180 degree shift was missing
in the conversion.